feat(scripts): finish partial dump script
